+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+- jrl_release: improve logging
+
 ## [2.2.0] - 2026-08-07
 
 - jrl_release: add `ConanFileVersionExtractor` by @arntanguy